Add isLoading global state; Rename all todo references to task; General cleanup;
